+++ src/Event/AbstractMailListener.php
@@ -3,27 +3,14 @@ declare(strict_types=1);
 
 namespace AcMailer\Event;
 
-use Zend\EventManager\AbstractListenerAggregate;
-use Zend\EventManager\EventManagerInterface;
-
 /**
  * Class AbstractMailListener
  * @author Alejandro Celaya Alastrué
  * @link http://www.alejandrocelaya.com
  */
-abstract class AbstractMailListener extends AbstractListenerAggregate implements MailListenerInterface
+abstract class AbstractMailListener implements MailListenerInterface
 {
-    /**
-     * @param EventManagerInterface $events
-     * @param int $priority
-     */
-    public function attach(EventManagerInterface $events, $priority = 1)
-    {
-        $this->listeners[] = $events->attach(MailEvent::EVENT_MAIL_PRE_RENDER, [$this, 'onPreRender'], $priority);
-        $this->listeners[] = $events->attach(MailEvent::EVENT_MAIL_PRE_SEND, [$this, 'onPreSend'], $priority);
-        $this->listeners[] = $events->attach(MailEvent::EVENT_MAIL_POST_SEND, [$this, 'onPostSend'], $priority);
-        $this->listeners[] = $events->attach(MailEvent::EVENT_MAIL_SEND_ERROR, [$this, 'onSendError'], $priority);
-    }
+    use MailListenerTrait;
 
     /**
      * Called before rendering the email, in case it is composed by a template